+++ src/container/IContainer.ts
@@ -25,8 +25,7 @@ export const TAGGED_PARAM = 'TAGGED_PARAM';
 // used to store class method to be injected
 export const TAGGED_METHOD = 'TAGGED_METHOD';
 
-// used to store router 
-export const CONTROLLER_ROUTER = "CONTROLLER_ROUTER";
+export type ClassOrString<T> = string | (new () => T);
 
 /**
  * defined AOP type
@@ -42,25 +41,11 @@ export enum AOPType {
 }
 
 /**
- * Aspect interface
- *
- * @export
- * @interface IAspect
- */
-export interface IAspect {
-  app: Application;
-
-  run: (...args: any[]) => Promise<any>;
-}
-
-/**
  * Defined constructor interface
  */
 export interface Constructor<T> {
   new(...args: any[]): T;
 }
-
-
 
 /**
  * Base Application interface
@@ -303,62 +288,18 @@ export interface ObjectDefinitionOptions {
   initMethod?: string;
   destroyMethod?: string;
   scope?: Scope;
-  type: ComponentType;
-  args: any[];
+  type?: ComponentType;
+  args?: any[];
 }
 
 /**
- *
- *
- * @export
- * @interface TagClsMetadata
- */
-export interface TagClsMetadata {
-  id: string;
-  originName: string;
-}
-
-/**
- *
+ * Aspect interface
  *
  * @export
- * @interface TagPropsMetadata
+ * @interface IAspect
  */
-export interface TagPropsMetadata {
-  key: string | number | symbol;
-  value: any;
-}
+export interface IAspect {
+  app: Application;
 
-/**
- *
- *
- * @export
- * @interface ReflectResult
- */
-export interface ReflectResult {
-  [key: string]: TagPropsMetadata[];
-}
-
-/**
- * Interface for Controller
- */
-export interface IController {
-  readonly app: Application;
-  readonly ctx: Context;
-}
-
-/**
- * Interface for Middleware
- */
-export interface IMiddleware {
-  run: (options: object, app: Application) => (ctx: object, next: Function) => Promise<any>;
-}
-
-/**
- * Interface for Service
- */
-export interface IService {
-  readonly app: Application;
-
-  // init(...arg: any[]): void;
+  run: (...args: any[]) => Promise<any>;
 }