+++ test/ClassC.ts
@@ -3,13 +3,14 @@
  * @Usage: 
  * @Author: richen
  * @Date: 2024-11-05 10:35:23
- * @LastEditTime: 2024-11-05 23:45:00
+ * @LastEditTime: 2025-02-26 17:46:01
  * @License: BSD (3-Clause)
  * @Copyright (c): <richenlin(at)gmail.com>
  */
 
-import { After, Before } from "../src/AOP";
+import { After, Before } from "../src/decorator/AOP";
 import { ClassA } from "./ClassA";
+import { TestAspect } from "./TestAspect";
 
 export class ClassC extends ClassA {
   name: string;
@@ -30,12 +31,12 @@ export class ClassC extends ClassA {
     return this.myDependency.run();
   }
 
-  @Before("TestAspect")
+  @Before(TestAspect)
   run2(event: string) {
     return true;
   }
 
-  @After("TestAspect")
+  @After(TestAspect)
   run3(event: string) {
     return true;
   }